Add Alpine Linux Containerfile
In the future we want to isolate the backup
process into its own container. For that we
need a small container that provides a shell
and has the `restic` binary readily available.
I have chosen to use Alpine Linux [1] for this purpose.
Distroless containers containers [2] were ruled
out, because I want to use interactive commands
such as `restic restore` and `restic snapshots`
inside of a dedicated shell.

This image makes use of `doas` instead of `sudo`, as
`doas` is the new default temporary privilege escalation tool
on Alpine Linux [3].

[1] https://www.alpinelinux.org/
[2] https://github.com/GoogleContainerTools/distroless
[3] https://wiki.alpinelinux.org/wiki/Release_Notes_for_Alpine_3.15.0#Move_from_sudo_to_doas