Add `--retry-lock=1h` flag to `restic` calls
Currenty every `restic` operation may fail if the restic
repositoy is locked. It might therefore be a good idea, to
let restic retry that operation at a later time. Here we
use the `--retry-lock=1h` option, to let restic retry that
operation every 10 seconds [1]. If it was not able to create a
lock within 1 hour, it will eventually fail.

[1] https://github.com/jooola/restic/commit/aade8379df5a36d9ff2982e81d3487cf354699f0